refactor: extract params
