+++ packages/functions/src/sdk.ts
@@ -2,4 +2,5 @@ export * from './sdk/controllers.sdk';
 export * from './sdk/db.sdk';
 export * from './sdk/schemas/controllers';
 export * from './sdk/schemas/db';
+export * from './sdk/schemas/params';
 export * from './sdk/serializer.sdk';
